refactore cli commands/groups to reduce __main__.py line count
