+++ seedfarmer/cli_groups/_remove_group.py
@@ -1,0 +1,70 @@
+#  Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+#
+#    Licensed under the Apache License, Version 2.0 (the "License").
+#    You may not use this file except in compliance with the License.
+#    You may obtain a copy of the License at
+#
+#        http://www.apache.org/licenses/LICENSE-2.0
+#
+#    Unless required by applicable law or agreed to in writing, software
+#    distributed under the License is distributed on an "AS IS" BASIS,
+#    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#    See the License for the specific language governing permissions and
+#    limitations under the License.
+
+import logging
+
+import click
+
+import seedfarmer.mgmt.module_info as mi
+from seedfarmer import DEBUG_LOGGING_FORMAT, enable_debug
+from seedfarmer.config import PROJECT
+
+_logger: logging.Logger = logging.getLogger(__name__)
+
+
+@click.group(name="remove", help="Top Level command to support removing module metadata")
+def remove() -> None:
+    f"""Remove module data for {PROJECT.upper()}"""
+    pass
+
+
+@remove.command(name="moduledata", help="Remove all SSM parameters tied to the module")
+@click.option(
+    "--deployment",
+    "-d",
+    type=str,
+    help="The Deployment Name",
+    required=True,
+)
+@click.option(
+    "--group",
+    "-g",
+    type=str,
+    help="The Group Name",
+    required=True,
+)
+@click.option(
+    "--module",
+    "-m",
+    type=str,
+    help="The Module Name",
+    required=True,
+)
+@click.option(
+    "--debug/--no-debug",
+    default=False,
+    help="Enable detailed logging.",
+    show_default=True,
+)
+def remove_module_data(
+    deployment: str,
+    group: str,
+    module: str,
+    debug: bool,
+) -> None:
+    if debug:
+        enable_debug(format=DEBUG_LOGGING_FORMAT)
+    _logger.debug(" We are removing module data for %s of group %s in %s", module, group, deployment)
+
+    mi.remove_module_info(deployment, group, module)
